Refactor: NCF
1. NCF modularization

NCF 모델 모듈화 #62
