+++ src/NCF/ncf.yaml
@@ -1,0 +1,16 @@
+data:
+  file_path: "data/train/train_ratings.csv"
+  output_path: "output/NCF.csv"
+
+model:
+  embed_dim: 16
+  hidden_dim: 64
+
+training:
+  epochs: 10
+  lr: 0.001
+  batch_size: 64
+  num_negatives: 4
+
+recommendation:
+  top_k: 10
